+++ src/interfaces/ISkillOptions.ts
@@ -1,0 +1,4 @@
+export default interface ISKillOptions {
+  event: any;
+  additionalMods?: (string | number)[];
+}