+++ src/module/settings.ts
@@ -69,4 +69,22 @@ export const registerSettings = function () {
     type: Number,
     config: true,
   });
+
+  game.settings.register('swade', 'vehicleMods', {
+    name: game.i18n.localize('SWADE.VehicleMods'),
+    hint: game.i18n.localize('SWADE.VehicleModsDesc'),
+    default: false,
+    scope: 'world',
+    type: Boolean,
+    config: true,
+  });
+
+  game.settings.register('swade', 'vehicleEdges', {
+    name: game.i18n.localize('SWADE.VehicleEdges'),
+    hint: game.i18n.localize('SWADE.VehicleEdgesDesc'),
+    default: false,
+    scope: 'world',
+    type: Boolean,
+    config: true,
+  });
 };