+++ src/module/sheets/SwadeNPCSheet.ts
@@ -1,9 +1,11 @@
 // eslint-disable-next-line no-unused-vars
-import { SwadeActor } from '../entities/SwadeActor';
+import SwadeActor from '../entities/SwadeActor';
 import * as chat from '../chat';
-import { SwadeBaseActorSheet } from './SwadeBaseActorSheet';
+import SwadeBaseActorSheet from './SwadeBaseActorSheet';
 
-export class SwadeNPCSheet extends SwadeBaseActorSheet {
+export default class SwadeNPCSheet extends SwadeBaseActorSheet {
+  actor: SwadeActor;
+
   static get defaultOptions() {
     return mergeObject(super.defaultOptions, {
       classes: ['swade', 'sheet', 'actor', 'npc'],
@@ -41,7 +43,8 @@ export class SwadeNPCSheet extends SwadeBaseActorSheet {
     // Resize resizable classes
     let resizable = (html as JQuery).find('.resizable');
     resizable.each((_, el) => {
-      let heightDelta = this.position.height - (this.options.height as number);
+      let heightDelta =
+        (this.position.height as number) - (this.options.height as number);
       el.style.height = `${heightDelta + parseInt(el.dataset.baseSize)}px`;
     });
     return html;
@@ -59,7 +62,12 @@ export class SwadeNPCSheet extends SwadeBaseActorSheet {
         li.setAttribute('draggable', 'true');
         li.addEventListener('dragstart', handler, false);
       });
-      html.find('div.item.weapon').each((i, li) => {
+      html.find('li.item.weapon').each((i, li) => {
+        // Add draggable attribute and dragstart listener.
+        li.setAttribute('draggable', 'true');
+        li.addEventListener('dragstart', handler, false);
+      });
+      html.find('li.item.power').each((i, li) => {
         // Add draggable attribute and dragstart listener.
         li.setAttribute('draggable', 'true');
         li.addEventListener('dragstart', handler, false);
@@ -85,17 +93,17 @@ export class SwadeNPCSheet extends SwadeBaseActorSheet {
 
     // Roll Skill
     html.find('.skill.item a').click((event) => {
-      let actorObject = this.actor as SwadeActor;
       let element = event.currentTarget as Element;
       let item = element.parentElement.dataset.itemId;
-      actorObject.rollSkill(item, { event: event });
+      this.actor.rollSkill(item, { event: event });
     });
 
     // Add new object
-    html.find('.item-create').click((event) => {
+    html.find('.item-create').click(async (event) => {
       event.preventDefault();
       const header = event.currentTarget;
       let type = header.dataset.type;
+      let createdItem: Item;
 
       // item creation helper func
       let createItem = function (
@@ -113,14 +121,18 @@ export class SwadeNPCSheet extends SwadeBaseActorSheet {
 
       // Getting back to main logic
       if (type == 'choice') {
-        this._chooseItemType().then((dialogInput: any) => {
-          const itemData = createItem(dialogInput.type, dialogInput.name);
-          this.actor.createOwnedItem(itemData, {});
+        this._chooseItemType().then(async (dialogInput: any) => {
+          let itemData = createItem(dialogInput.type, dialogInput.name);
+          itemData.data.equipped = true;
+          createdItem = await this.actor.createOwnedItem(itemData, {});
+          this.actor.getOwnedItem(createdItem._id).sheet.render(true);
         });
         return;
       } else {
-        const itemData = createItem(type);
-        this.actor.createOwnedItem(itemData, {});
+        let itemData = createItem(type);
+        itemData.data.equipped = true;
+        createdItem = await this.actor.createOwnedItem(itemData, {});
+        this.actor.getOwnedItem(createdItem._id).sheet.render(true);
       }
     });
 
@@ -160,12 +172,12 @@ export class SwadeNPCSheet extends SwadeBaseActorSheet {
         await this.actor.update({
           'data.details.conviction.active': false,
         });
-        chat.createConvictionEndMessage(this.actor as SwadeActor);
+        chat.createConvictionEndMessage(this.actor);
       }
     });
   }
 
-  getData() {
+  getData(): ActorSheetData {
     let data: any = super.getData();
 
     // Everything below here is only needed if user is not limited
@@ -211,7 +223,12 @@ export class SwadeNPCSheet extends SwadeBaseActorSheet {
       });
     }
 
-    data.armor = (this.actor as SwadeActor).calcArmor();
+    data.armor = this.actor.calcArmor();
+    if (
+      data.armor !== getProperty(this.actor.data, 'data.stats.toughness.armor')
+    ) {
+      this.actor.update({ 'data.stats.toughness.armor': data.armor });
+    }
 
     const shields = data.itemsByType['shield'];
     data.parry = 0;